Expose catalog items via CLI, reorganize commands
